add dtype selection and progess bar to bin_data_mmap
Former-commit-id: d031bbc1efa6f928f643f1f0966990d92d00a994

+++ py4DSTEM/io/datastructure/datacube.py
@@ -96,7 +96,7 @@ class DataCube(DataObject):
     def bin_data_diffraction(self, bin_factor):
         self = preprocess.bin_data_diffraction(self, bin_factor)
 
-    def bin_data_mmap(self, bin_factor):
+    def bin_data_mmap(self, bin_factor, dtype=np.float32):
         self = preprocess.bin_data_mmap(self, bin_factor)
 
     def bin_data_real(self, bin_factor):
